% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.ints.R
\name{conf.ints}
\alias{conf.ints}
\title{Helper function to construct confidence intervals}
\usage{
conf.ints(model = model, level = 0.95, vars = NULL,
  var.labels = NULL)
}
\arguments{
\item{model}{Model takes the object name of your linear regression model. Example - if you run the
following command: m1 <- lm(y~x), then enter m1 for this argument.}

\item{level}{User-specified confidence interval. Default = 0.95}

\item{vars}{Variable selection (default all)}

\item{var.labels}{Variable names if not default}
}
\description{
Helper function that builds out user-specified confidence intervals.
}
\examples{
# Estimate a regression using 'mtcars' data:
m <- lm(mpg~disp+hp+wt+am,data=mtcars)
conf.ints(m,level=0.99)

}
