% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R
\name{confidence}
\alias{confidence}
\title{Plot confidence intervals for variables from your linear regression model output}
\usage{
confidence(model, vars = NULL, var.labels = NULL, level = 0.95,
  grid = T, grid.int = 5, mai = NULL, ps = 12, pt.type = 16,
  pt.size = 1.5, pt.color = "red", lwd = 1.5, lty = 2,
  legend = F)
}
\arguments{
\item{model}{Model takes the object name of your linear regression model. Example - if you run the
following command: m1 <- lm(y~x), then enter m1 for this argument.}

\item{var.labels}{New variable names for the plot to be entered as characters here.
Note, this must be the same length as the variables in your model. If you have four variables, you must
enter four names. See example code below.}

\item{grid}{Include vertical grid lines in the plot. Default is True.}

\item{grid.int}{Spacing interval for grid lines. Default value is 5.}

\item{mai}{Change plot margins (inches).}

\item{ps}{Change text size in plot. This will be useful for your posters
(30 would be good for a poster, 12 for a paper).}

\item{pt.type}{Point style of regression parameter point estimate. Default is 16, circle. See ?points
for information on different styles.}

\item{pt.size}{Side of point estimate, default is 1.5.}

\item{pt.color}{Color of point estimate, default is 'red', Any color argument will work here.}

\item{lwd}{Line width for confidence intervals, default is 1.5.}

\item{lty}{Line type for confidence intervals, default is 2 (dashed line). See the lty argument
in ?par for additional details on line types.}
}
\description{
This function takes the linear regression model you estimate with lm() for your final
projects and returns a coefficient plot appropriate for you to use in your poster or final paper.
}
\examples{
# Estimate a regression using 'mtcars' data:
m <- lm(mpg~disp+hp+wt,data=mtcars)
confidence(m)

# Plot without intercept:
confidence(m,vars=c('disp','hp','wt'))

# Plot with intercept and one covariate:
confidence(m,vars=c('(Intercept)','wt'))

# Plot with two model variables AND new variable names:
confidence(m,vars=c('disp','wt'),var.labels=c('Displacement','Weight'))

# Plot with higher confidence interval and legend:
confidence(m,level=0.99,legend=TRUE)

# Dynamic margins for longer variable names:
confidence(m, vars=c('hp','wt'),var.labels=c('A really long name for variable: Horse Power','Weight'))
}
