% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{effect}
\alias{effect}
\title{Plot effect for one variable from your linear regression model output}
\usage{
effect(model, var, level = 0.95, dummyIV = F, x.label = NULL,
  y.label = NULL, title = NULL, grid = F, rug = F)
}
\arguments{
\item{model}{The object name of your linear regression model. Example - if you run the
following command: model1 <- lm(y~x), then enter 'model1' for this argument.}

\item{var}{The variable to plot change in predicted values of 'y.' Your 'x' variable.}

\item{level}{Confidence level, .95 is the default.}

\item{dummyIV}{Indicator for whether the plotted IV is a dummy variable, default is FALSE}

\item{x.label}{Option to change x-axis label in plot}

\item{y.label}{Option to change y-axis label in plot}

\item{title}{TO FINISH -- option to add plot title}

\item{grid}{Option to plot with grid lines, default is FALSE}

\item{rug}{Option to plot with rug to represent data density in plot, default is FALSE}

\item{color}{TO FINISH -- option to change polygon colors}

\item{line.width}{TO FINISH -- option to change fit regression line thickness}
}
\description{
This function takes the linear regression model you estimate with lm() for your final
projects and returns a plot of an effect for a selected variable.
}
\examples{
# Estimate a regression using 'mtcars' data:
m <- lm(mpg~disp+hp+wt+am,data=mtcars)

# Plot effect of 'wt' variable
effect(m,'wt')

# Plot effect of dummy variable, 'am'
effect(m,'am',dummyIV=TRUE)

# Plot with user-defined x-axis label
effect(m,'wt',x.label='Vehicle Weight')

# Plot with grid lines and rug
effect(m,'wt',x.label='Vehicle Weight', grid=TRUE,rug=TRUE)
}
