% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.mat.R
\name{mean.mat}
\alias{mean.mat}
\title{Helper function to construct matrix for effects plot}
\usage{
\method{mean}{mat}(model, var)
}
\arguments{
\item{model}{Model takes the object name of your linear regression model. Example - if you run the
following command: m1 <- lm(y~x), then enter m1 for this argument.}

\item{var}{The variable to plot change in predicted values of 'y.' Your 'x' variable. Will inheret from
parent function except when called independently.}
}
\description{
Helper function that constructs a matrix of select variable with all others held at their means. Will inheret
call options from parent function unless called independently.
}
\examples{
#' # Estimate a regression using 'mtcars' data:
m <- lm(mpg~disp+hp+wt+am,data=mtcars)
mean.mat(m,var='wt')
}
